# src/search_service/models/bad_request.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class BadRequest(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    BadRequest - a model defined in OpenAPI

        message: The message of this BadRequest [Optional].
    """

    message: Optional[str] = None

BadRequest.update_forward_refs()