# src/search_service/models/recommend_context.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class RecommendContext(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    RecommendContext - a model defined in OpenAPI

        user_uid: The user_uid of this RecommendContext.
        visit_id: The visit_id of this RecommendContext [Optional].
        page_id: The page_id of this RecommendContext [Optional].
        panel_id: The panel_id of this RecommendContext [Optional].
    """

    user_uid: str
    visit_id: Optional[str] = None
    page_id: Optional[str] = None
    panel_id: Optional[str] = None

RecommendContext.update_forward_refs()