# src/search_service/models/recommend_results.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401
from search_service.models.search_result import SearchResult


class RecommendResults(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    RecommendResults - a model defined in OpenAPI

        results: The results of this RecommendResults.
    """

    results: List[SearchResult]

RecommendResults.update_forward_refs()