# src/search_service/models/search_result.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class SearchResult(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    SearchResult - a model defined in OpenAPI

        collection: The collection of this SearchResult.
        id: The id of this SearchResult.
        attributes: The attributes of this SearchResult [Optional].
    """

    collection: str
    id: str
    attributes: Optional[Dict[str, Any]] = None

SearchResult.update_forward_refs()